Refactor Args class to libs.domain.datamodels for improved organization and clarity
