+++ libs/domain/datamodels.py
@@ -13,9 +13,47 @@ from libs.utils.timekit import ExtendedDatetime as ExtDt
 from libs.utils.timekit import Format
 
 if TYPE_CHECKING:
+    from pathlib import Path
+
     from integrations.protocols import MessageParserProtocol
     from libs.domain.score import GameResult
     from libs.types import RemarkDict
+
+
+@dataclass
+class Args:
+    """コマンドラインオプション"""
+
+    service: str
+    config: "Path"
+    """設定ファイルパス"""
+
+    debug: int
+    """デバッグ出力フラグ"""
+    verbose: int
+    """詳細出力フラグ"""
+
+    moderate: bool
+    """INFO以下のログレベル出力を抑止"""
+    notime: bool
+    """ログに日付を付与しない"""
+
+    # Only allowed when --service=standard_io
+    text: str
+
+    # Only allowed when --service=web
+    host: str
+    port: int
+
+    # dbtools
+    compar: bool
+    unification: "Path"
+    recalculation: bool
+    export_data: str
+    import_data: str
+    vacuum: bool
+    gen_test_data: int
+    testcase: Optional["Path"]
 
 
 @dataclass
