+++ libs/global_value.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
     from integrations.standard_io.adapter import ServiceAdapter as std_adapter
     from integrations.web.adapter import ServiceAdapter as web_adapter
     from libs.bootstrap.app_config import AppConfig
-    from libs.types import Args
+    from libs.domain.datamodels import Args
 
 # --- グローバル変数 ---
 adapter: Union["slack_adapter", "discord_adapter", "web_adapter", "std_adapter"]
